refactor(vault): move create vault validations to new file
